Change long to integer

+++ Application/Features/Account/Commands/AddUserRole/AddUserRoleViewModel.cs
@@ -2,9 +2,9 @@ namespace Application.Features.Account.Commands.AddUserRole
 {
     public class AddUserRoleViewModel
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public string UserName { get; set; }
-        public long RoleId { get; set; }
+        public int RoleId { get; set; }
         public string RoleName { get; set; }
     }
 }
