+++ Application/Features/Documents/Commands/AssignUserPermission/AssignUserPermissionViewModel.cs
@@ -2,9 +2,9 @@ namespace Application.Features.Documents.Commands.AssignUserPermission
 {
     public class AssignUserPermissionViewModel
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public string UserName { get; set; }
-        public long DocumentId { get; set; }
+        public int DocumentId { get; set; }
         public string DocumentName { get; set; }
     }
 }
