+++ Domain/Entities/GroupDocument.cs
@@ -2,9 +2,9 @@ namespace Domain.Entities
 {
     public class GroupDocument
     {
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
         public string GroupName { get; set; }
-        public long DocumentId { get; set; }
+        public int DocumentId { get; set; }
         public string DocumentName { get; set; }
     }
 }
