+++ Domain/Entities/Role.cs
@@ -3,11 +3,11 @@ using Microsoft.AspNetCore.Identity;
 
 namespace Domain.Entities
 {
-    public class Role : IdentityRole<long>
+    public class Role : IdentityRole<int>
     {
-        public virtual long CreatedBy { get; set; }
+        public virtual int CreatedBy { get; set; }
         public virtual DateTime CreatedAt { get; set; }
-        public virtual long UpdatedBy { get; set; }
+        public virtual int UpdatedBy { get; set; }
         public virtual DateTime UpdatedAt { get; set; }
     }
 }
