+++ Domain/Entities/UserGroup.cs
@@ -2,10 +2,10 @@ namespace Domain.Entities
 {
     public class UserGroup
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public User User { get; set; }
 
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
         public Group Group { get; set; }
     }
 }
